train.py
import logging
import os
from typing import Dict, Tuple

import torch
import torch.nn as nn
import torch.optim as optim
import torchvision
import torchvision.transforms as transforms
import yaml
from tqdm import tqdm

# Load configuration
with open("config.yaml", "r") as f:
    config = yaml.safe_load(f)

# Set up logging
logging.basicConfig(
    level=config["logging"]["level"],
    format=config["logging"]["format"],
    handlers=[
        logging.FileHandler(config["logging"]["file"]),
        logging.StreamHandler(),
    ],
)
logger = logging.getLogger(__name__)

# Type hints
Dataset = Tuple[torch.utils.data.Dataset, torch.utils.data.Dataset, torch.utils.data.Dataset]
ModelOutput = Tuple[nn.Module, optim.Optimizer, nn.CrossEntropyLoss]


def download_data() -> None:
    """Download and extract the dataset."""
    # Download and extract the dataset (implementation omitted for brevity)
    pass


def load_data() -> Dataset:
    """Load and preprocess the dataset."""
    # Load and preprocess the dataset (implementation omitted for brevity)
    return train_dataset, val_dataset, test_dataset


def create_model() -> ModelOutput:
    """Create the model, optimizer, and loss function."""
    # Create the model
    model = getattr(torchvision.models, config["model"]["name"])(
        pretrained=config["model"]["pretrained"],
        num_classes=config["model"]["num_classes"],
    )

    # Freeze base layers if specified
    if config["model"]["freeze_base"]:
        for param in model.parameters():
            param.requires_grad = False

    # Create the optimizer
    optimizer_config = config["training"]["optimizer"]
    optimizer = getattr(optim, optimizer_config["name"])(
        model.parameters(), lr=optimizer_config["lr"]
    )

    # Create the loss function
    loss_fn = getattr(nn, config["training"]["loss"] + "Loss")()

    return model, optimizer, loss_fn


def train(
    model: nn.Module, optimizer: optim.Optimizer, loss_fn: nn.CrossEntropyLoss, train_loader: torch.utils.data.DataLoader, val_loader: torch.utils.data.DataLoader
) -> Dict[str, float]:
    """Train the model and evaluate on the validation set."""
    # Training loop (implementation omitted for brevity)
    return {"accuracy": 0.9}  # Example output


def main() -> None:
    # Download the dataset
    download_data()

    # Load and preprocess the dataset
    train_dataset, val_dataset, test_dataset = load_data()

    # Create the model, optimizer, and loss function
    model, optimizer, loss_fn = create_model()

    # Create data loaders
    train_loader = torch.utils.data.DataLoader(train_dataset, batch_size=config["training"]["batch_size"], shuffle=True)
    val_loader = torch.utils.data.DataLoader(val_dataset, batch_size=config["training"]["batch_size"])

    # Train the model
    metrics = train(model, optimizer, loss_fn, train_loader, val_loader)
    logger.info(f"Validation metrics: {metrics}")

    # Save artifacts
    os.makedirs(config["artifacts"]["output_dir"], exist_ok=True)
    if config["artifacts"]["save_best_model"]:
        torch.save(model.state_dict(), os.path.join(config["artifacts"]["output_dir"], "best_model.pth"))
    if config["artifacts"]["save_logs"]:
        # Save logs (implementation omitted for brevity)
        pass
    if config["artifacts"]["save_metrics"]:
        # Save metrics (implementation omitted for brevity)
        pass


if __name__ == "__main__":
    main()