pyproject.toml
[tool.poetry]
name = "lb1_tml"
version = "0.1.0"
description = "lab 1 texn. mashine learning"
authors = ["Kateryna Barbasheva <katherina.barbasheva@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.10.0"
torchvision = "^0.11.1"
tqdm = "^4.64.1"
matplotlib = "^3.6.3"
numpy = "^1.21.6"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
ruff = "^0.0.254"
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"